third_party/blink/web_tests/platform/generic/fast/css/invalidation/empty-pseudo-expected.txt

Use descendant invalidation set for :empty pseudo class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(empty, '').backgroundColor is green
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(empty, '').backgroundColor is transparent
PASS getComputedStyle(notEmpty, '').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(notEmpty, '').backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
